requirements.txt
fastapi
uvicorn[standard]
pydantic_settings
SQLAlchemy
pymysql